+++ docs/conf.py
@@ -18,7 +18,7 @@
 # -- Project information -----------------------------------------------------
 
 project = 'pyigt'
-copyright = '2022, Johann-Mattis List and Robert Forkel'
+copyright = '2023, Johann-Mattis List and Robert Forkel'
 author = 'Johann-Mattis List and Robert Forkel'
 
 # The full version, including alpha/beta/rc tags
@@ -56,7 +56,7 @@ html_theme = 'sphinx_rtd_theme'
 # Add any paths that contain custom static files (such as style sheets) here,
 # relative to this directory. They are copied after the builtin static files,
 # so a file named "default.css" will overwrite the builtin "default.css".
-html_static_path = ['_static']
+html_static_path = []
 
 typehints_fully_qualified = True
 always_document_param_types = True